# public/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Online Shopper",
        "contact": {
            "name": "E4 Class G17 Group ?",
            "url": "https://www.kuylim.wordpress.com"
        },
        "version": "1.0"
    },
    "host": "online-shop.com",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/products/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "description": "Returns a product based on a single ID",
                "operationId": "findProductById",
                "produces": [
                    "application/json",
                    "application/xml",
                    "text/html",
                    "text/xml"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of product to fetch",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "product found"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Product"
                ],
                "description": "Update old product.",
                "operationId": "updateProduct",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of product to update",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "user",
                        "in": "body",
                        "description": "Product to update to the database",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Save Product Success"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "description": "deletes a single product based on the product id",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of product to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "product deleted success"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            }
        },
...
        "/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Return all users",
                "operationId": "finadAllUser",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "User found"
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "description": "Creates a new user.",
                "operationId": "addUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "description": "User to add to the Database",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Save User Success"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "Returns a user based on a single ID",
                "operationId": "findUserByID",
                "produces": [
                    "application/json",
                    "application/xml",
                    "text/html",
                    "text/xml"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user to fetch",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user found"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "description": "Update a old user.",
                "operationId": "updateProduct",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user to update",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "user",
                        "in": "body",
                        "description": "User to update to the database",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Save User Success"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "description": "deletes a single user based on the user id",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user deleted success"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            }
        },
        "/users/search/{email}": {
            "get": {
                "tags": [
                    "User"
                ],
                "description": "Returns a user based on a email",
                "operationId": "findUserByEmail",
                "produces": [
                    "application/json",
                    "application/xml",
                    "text/html",
                    "text/xml"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "Email of user to fetch",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user found"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            }
        }
    },
...
}